Position Code

+++ FRC5012-Code-2017/src/com/gryffingear/y2017/utilities/GryffinMath.java
@@ -38,7 +38,7 @@ public class GryffinMath {
 	 *            output value's upper limit
 	 * @return the scaled value
 	 */
-	public static float map(float value, float istart, float istop, float ostart, float ostop) {
+	public static double map(double value, double istart, double istop, double ostart, double ostop) {
 
 		return ostart + (ostop - ostart) * ((value - istart) / (istop - istart));
 	}
